added countless features

+++ Program.cs
@@ -1,6 +1,4 @@
 ﻿using FFMpegCore;
-using FFMpegCore.Enums;
-using System.IO;
 
 bool Verbose = true;
 
@@ -11,7 +9,7 @@ if(Verbose)
     Console.ForegroundColor = ConsoleColor.DarkYellow;
     Console.WriteLine("#########################################");
     Console.WriteLine("##                                     ##");
-    Console.WriteLine("##    =====    FFCONVERT    =====      ##");
+    Console.WriteLine("##     =====      FFBOT      =====     ##");
     Console.WriteLine("##                                     ##");
     Console.WriteLine("#########################################");
     Console.ResetColor();
@@ -21,15 +19,9 @@ if(Verbose)
 HashSet<string> SkipBuffer = new();
 
 string WorkingDirectoryPath = Environment.CurrentDirectory;
-string TempDirectoryPath = Path.GetTempPath();
-if(Verbose)
-{
-    Console.WriteLine($"Working directory: {WorkingDirectoryPath}");
-    Console.WriteLine($"Temp directory: {TempDirectoryPath}");
-}
+string? ForcedDestinationDirectoryPath = "../Encoded";
 
-
-string[] AllowedInputs = new[] { ".mkv", ".avi", ".vp9", ".ts" };
+string[] AllowedInputs = new[] { ".mkv", ".avi", ".vp9", ".ts", ".mp4", ".webm" };
 string OutputExtension = "mkv";
 string TargetedVideoCodec = "vp9";
 string? BaseCustomInputArguments = "-y -probesize 1000000000 -analyzeduration 100000000";
@@ -37,27 +29,56 @@ string? BaseCustomOutputArguments = "";
 bool ShouldSetRatioToOneOne = true;
 bool SmartAudioEncoding = true;
 bool KeepOnlyOneVideoStream = true;
-if(Verbose)
+bool ShouldRemoveOldFile = true;
+bool ShouldSendRemovedToBin = true;
+bool ShouldKillFFMpegWhenExited = true;
+bool ShouldDeleteEmptyDirectories = true;
+
+TimeSpan WaitingTime = TimeSpan.FromSeconds(60);
+
+if (Verbose)
 {
-    Console.WriteLine($"Should set aspect ratio to 1:1: {ShouldSetRatioToOneOne}");
+    Console.WriteLine($"Working directory: {WorkingDirectoryPath}");
+    if(ForcedDestinationDirectoryPath != null)
+        Console.WriteLine($"Forced a destination path : {ForcedDestinationDirectoryPath}");
+
     Console.WriteLine($"Allowed inputs: {string.Join(",", AllowedInputs)}");
-    Console.WriteLine($"Base custom input arguments: {BaseCustomInputArguments}");
     Console.WriteLine($"Output extension: {OutputExtension}");
+    Console.WriteLine($"Targeted video codec: {TargetedVideoCodec}");
+
+    Console.WriteLine($"Base custom input arguments: {BaseCustomInputArguments}");
     Console.WriteLine($"Base custom input arguments: {BaseCustomOutputArguments}");
+    Console.WriteLine($"Should set aspect ratio to 1:1: {ShouldSetRatioToOneOne}");
+    Console.WriteLine($"Should use smart audio encoding: {SmartAudioEncoding}");
+    Console.WriteLine($"Should keep only one video stream: {KeepOnlyOneVideoStream}");
+
+    Console.WriteLine($"Idle time: {WaitingTime}");
+
     Console.Write(Environment.NewLine);
 }
 
-TimeSpan WaitingTime = TimeSpan.FromSeconds(60);
+if(ForcedDestinationDirectoryPath != null && !Directory.Exists(ForcedDestinationDirectoryPath))
+    Directory.CreateDirectory(ForcedDestinationDirectoryPath);
+
+Action CancelCurrentFFMPeg = () => { };
+AppDomain.CurrentDomain.ProcessExit += new EventHandler(OnExited);
+
+if (!await Handle())
+{
+    Console.WriteLine($"Ready to encode future files put in \"{WorkingDirectoryPath}\"");
+    await Task.Delay( WaitingTime );
+}
 while (true)
 {
     if (!await Handle())
-        await Task.Delay((int)WaitingTime.TotalMilliseconds);
+        await Task.Delay(WaitingTime);
 }
 
 //return true if convert, false otherwise
 async Task<bool> Handle()
 {
-    var candidates = Directory.GetFiles(WorkingDirectoryPath, "*", searchOption: SearchOption.AllDirectories)
+    var rawCandidates = Directory.GetFiles(WorkingDirectoryPath, "*", searchOption: SearchOption.AllDirectories);
+    var candidates = rawCandidates
         .Where(file => AllowedInputs.Any(file.ToLower().EndsWith))
         .ToList();
 
@@ -76,7 +97,17 @@ async Task<bool> Handle()
             continue;
         }
 
-        var mediaInfo = await FFProbe.AnalyseAsync(InputPath);
+        IMediaAnalysis? mediaInfo = null;
+        try
+        {
+            mediaInfo = await FFProbe.AnalyseAsync(InputPath);
+        }
+        catch(Exception e)
+        {
+            SkipBuffer.Add(InputPath);
+            Console.WriteLine($"Cannot analyse {InputPath}: {e}");
+            continue;
+        }
         var totalTime = mediaInfo.Duration;
 
         //Skip already converted files
@@ -126,13 +157,13 @@ async Task<bool> Handle()
                 }
                 else
                     CustomOutputArgs += $" -c:s:{i} copy";
-
-                if (hasTextSubtitle)
-                    CustomInputsArgs += $" -txt_format text -fix_sub_duration";
             }
+            if (hasTextSubtitle)
+                CustomInputsArgs += $" -txt_format text -fix_sub_duration";
         }
 
-        Console.WriteLine($"Converting {InputPath}...");
+        if(Verbose)
+            Console.WriteLine($"Converting {InputPath}...");
 
         string OutputPath = Path.Combine(WorkingDirectoryPath, "temporary" + "." + OutputExtension);
         if (File.Exists(OutputPath))
@@ -145,6 +176,7 @@ async Task<bool> Handle()
             .OutputToFile(OutputPath, false, options => options
                 .WithCustomArgument(CustomOutputArgs)
             )
+            .CancellableThrough(out CancelCurrentFFMPeg);
             /*.NotifyOnError((string message) =>
             {
                 Console.Write($"{message}");
@@ -154,15 +186,29 @@ async Task<bool> Handle()
             {
                 Console.WriteLine($"Message: {message}");
             })*/
-            .NotifyOnProgress((double x) =>
-            {
-                Console.Write((x).ToString("0.0") + "%");
-                Console.CursorLeft = 0;
-            }, totalTime);
             
-        Console.WriteLine($"Executed command: {ffmpegArgs.Arguments}");
+
+        if (Verbose)
+            Console.WriteLine($"Executed command: {ffmpegArgs.Arguments}");
+
+        string relativeInputPath = Path.GetRelativePath(WorkingDirectoryPath, InputPath);
+        string? InputDirRelative = null;
+        if(fiInput.DirectoryName != null)
+            InputDirRelative = Path.GetRelativePath(WorkingDirectoryPath, fiInput.DirectoryName);
+
+        var progress = new ProgressBar(suffix: $" {relativeInputPath}");
+
+        ffmpegArgs = ffmpegArgs.NotifyOnProgress((TimeSpan current) =>
+        {
+            var relativeProgress = current / totalTime;
+
+            progress.Report(relativeProgress);
+        });
 
         bool worked = await ffmpegArgs.ProcessAsynchronously(false);
+        CancelCurrentFFMPeg = () => { };
+        progress.Dispose();
+
         SkipBuffer.Add(InputPath);
         if (!worked)
         {
@@ -170,15 +216,45 @@ async Task<bool> Handle()
             SkipBuffer.Add(InputPath);
             return true;
         }
-            
+
+        var newMediaInfo = await FFProbe.AnalyseAsync(OutputPath);
+        if (ShouldEncode(newMediaInfo, true))
+            Console.WriteLine($"Attention, l'analyse du fichier {OutputPath} indique que le fichier reste à encoder.");
 
         if (fiInput.DirectoryName == null)
             throw new Exception($"Safe lock: empty DirectoryName for {fiInput}.");
 
-        string newInputPath = Path.Combine(fiInput.DirectoryName, Path.GetFileNameWithoutExtension(InputPath) + "." + OutputExtension);
+        string newInputPath;
+        string newInputFileName = Path.GetFileNameWithoutExtension(InputPath) + "." + OutputExtension;
+        if (ForcedDestinationDirectoryPath != null)
+            newInputPath = Path.Combine(ForcedDestinationDirectoryPath, InputDirRelative, newInputFileName);
+        else
+            newInputPath = Path.Combine(fiInput.DirectoryName, newInputFileName);
 
-        File.Delete(InputPath);
+        FileInfo newInputFi = new(newInputPath);
+        if(newInputFi.DirectoryName != null && !Directory.Exists(newInputFi.DirectoryName))
+            Directory.CreateDirectory(newInputFi.DirectoryName);
+
+        if(ShouldRemoveOldFile)
+        {
+                /*Microsoft.VisualBasic.FileIO.FileSystem.DeleteFile(
+                    InputPath,
+                    Microsoft.VisualBasic.FileIO.UIOption.OnlyErrorDialogs,
+                    ShouldSendRemovedToBin 
+                        ? Microsoft.VisualBasic.FileIO.RecycleOption.SendToRecycleBin 
+                        : Microsoft.VisualBasic.FileIO.RecycleOption.DeletePermanently);*/
+            File.Delete(InputPath);
+        }
+        
         File.Move(OutputPath, newInputPath);
+        if (!File.Exists(newInputPath))
+            Console.WriteLine($"Echec, aucun fichier à l'emplacement {newInputPath}.");
+
+        if (ShouldDeleteEmptyDirectories
+        && ForcedDestinationDirectoryPath != null
+        && fiInput.DirectoryName != null
+        && !Directory.EnumerateFileSystemEntries(fiInput.DirectoryName).Any())
+            Directory.Delete(fiInput.DirectoryName);
 
         Console.WriteLine($"Fichier remplacé avec succès: {InputPath}.");
         return true;
@@ -187,7 +263,7 @@ async Task<bool> Handle()
     return false;
 }
 
-bool ShouldEncode(FFMpegCore.IMediaAnalysis? media)
+bool ShouldEncode(FFMpegCore.IMediaAnalysis? media, bool verbose = false)
 {
     // skip bad analysis
     if (media == null)
@@ -199,14 +275,23 @@ bool ShouldEncode(FFMpegCore.IMediaAnalysis? media)
 
     // encode if more than 1 video stream
     if (KeepOnlyOneVideoStream && media.VideoStreams.Count > 1)
+    {
+        if (verbose) Console.WriteLine($"Le fichier a {media.VideoStreams.Count} flux videos.");
         return true;
+    }        
 
     // encode bad codec
     if (media.PrimaryVideoStream.CodecName != TargetedVideoCodec)
+    {
+        if (verbose) Console.WriteLine($"Le codec video du fichier est {media.PrimaryVideoStream.CodecName}.");
         return true;
+    }
 
     if (ShouldSetRatioToOneOne && IsBadAspectRatio(media))
+    {
+        if (verbose) Console.WriteLine($"Le fichier a un SampleAspectRatio de {media.PrimaryVideoStream.SampleAspectRatio.Width}:{media.PrimaryVideoStream.SampleAspectRatio.Height}.");
         return true;
+    }
 
     return false;
 }
@@ -221,4 +306,10 @@ bool IsBadAspectRatio(FFMpegCore.IMediaAnalysis media)
         return true;
 
     return false;
+}
+
+void OnExited(object? sender, EventArgs e)
+{
+    if (ShouldKillFFMpegWhenExited)
+        CancelCurrentFFMPeg();
 }
